+++ openfe/setup/transformation_factory/__init__.py
@@ -1,0 +1,1 @@
+from .easy_transformation_factory import EasyTransformationFactory
