+++ openfe/setup/transformation_factory/easy_transformation_factory.py
@@ -1,0 +1,44 @@
+# This code is part of OpenFE and is licensed under the MIT license.
+# For details, see https://github.com/OpenFreeEnergy/openfe
+
+from typing import FrozenSet, Dict
+
+from gufe import AlchemicalNetwork, Protocol, LigandAtomMapping, Transformation, SmallMoleculeComponent
+from gufe import Transformation
+
+from .abstract_transformation_factory import AbstractTransformerFactory
+
+class AbstractEasyTransformationFactory(AbstractTransformerFactory):
+
+    def __init__(self, protocol: Protocol, chemical_system_generator: callable):
+        self.protocol = protocol
+        self.chemical_system_generator = chemical_system_generator
+
+
+class RFETransformationFactory(AbstractEasyTransformationFactory):
+    """
+        Simple transformer is a very simplistic transformer, that mapps a defined protocol to a given set of transformations
+    """
+
+    def __call__(self, alchemical_network_edges: FrozenSet[LigandAtomMapping], name="approach") \
+            -> AlchemicalNetwork:
+        name = name
+        edges = []
+        nodes = []
+
+        for edge in alchemical_network_edges:
+
+            for stateA_env, stateB_env in zip(self.chemical_system_generator(edge.componentA), self.chemical_system_generator(edge.componentB)):
+                print(stateA_env)
+                transformation_name = name + "_" + stateA_env.name + "_" + stateB_env.name
+
+                edges.append(Transformation(stateA=stateA_env,
+                                            stateB=stateB_env,
+                                            mapping=edge,
+                                            name=transformation_name,
+                                            protocol=self.protocol)
+                             )
+                nodes.extend([stateA_env, stateB_env])
+
+        alchemical_network = AlchemicalNetwork(nodes=nodes, edges=edges, name=name)
+        return alchemical_network
